* 'feature/david-wip' of https://github.com/mcgill-robotics/rover: (30 commits)
  Fix pin
  Change for omnicam device;
  Change arm pins
  Convert arduino reset to object and add GPIO calls
  Add service to reset the arduino from the RPi GPIO, not complete
  Add arm camera to udev
  Change service stop behaviour;
  Do not return new virtual ahrs if not found
  Change master of init script
  Set proper camera names
  Change path for the rpi
  Do not respawn launch files
  Add auto-launch scripts for the rpi
  New launch files for rpi
  New launch files for rpi
  Moved
  Add arm tf boradcaster for desired joint positions
  Fix build
  Add scaling factor and use pot class in main
  Change encoder interface to add possible scaling factor for base yaw. Also account for possible multiple rotations of the encoder. Add potentiometer class for base pitch pot.
  ...

# Conflicts:
#	catkin_ws/src/arduino/CMakeLists.txt
#	catkin_ws/src/arduino/src/utilities/CMakeLists.txt

+++ catkin_ws/src/arduino/src/MotorController/src/DRV8308.h
@@ -13,8 +13,8 @@ public:
     DRV8308() { };
     virtual ~DRV8308() { };
 
-    virtual void setSpeed(int);
-    virtual void setDirection(int);
+    virtual void setSpeed(float);
+    virtual void setDirection(float);
     virtual void brake(bool);
     virtual void enable(bool);
     virtual bool getStatus();
